feat: small refactor for the search package
