+++ plugins/search-react/src/alpha/blueprints/SearchResultListItemBlueprint.test.tsx
