+++ plugins/search-react/src/alpha/blueprints/index.ts
