+++ plugins/search-react/src/alpha/blueprints/types.ts
